FX-3316 Refactoring
